+++ frontend/lib/main.dart
@@ -1,5 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:mybrary/screens/sign_up_screen.dart';
+import 'package:flutter/services.dart';
+import 'package:mybrary/screens/auth/login_screen.dart';
+import 'package:mybrary/screens/auth/sign_up_screen.dart';
+import 'package:mybrary/screens/auth/sign_up_verify_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -10,8 +13,21 @@ class MyApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: SignUpScreen(),
+    // System Interface 글자 색상 light, dark
+    // 현재는 전체 적용, 추후 일부 적용 예정
+    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle.dark);
+    return MaterialApp(
+      theme: ThemeData(
+        appBarTheme: AppBarTheme(
+          systemOverlayStyle: SystemUiOverlayStyle.dark,
+        ),
+      ),
+      initialRoute: '/',
+      routes: {
+        '/': (context) => LoginScreen(),
+        '/signup': (context) => SignUpScreen(),
+        '/signup/verify': (context) => SignUpVerifyScreen(),
+      },
     );
   }
 }
